Remove unnecessary animation from MyVideoNav
The unwatched episode count in Landscape, Showcase and Wall views is
collapsed on top of the NextAired label below. Previously, an animation
would push this label down if the genre was not displayed. However, this
animation is unneeded because nothing is hidden above the label when it
is in the lowered position. Removing the animation improves the
presentation of the label when NextAired information is shown.